+++ packages/mobile/locales/en-US/global.json
@@ -159,5 +159,6 @@
   "addressValidationFullOwnAddress": "Please enter your contact's account number, not your own. Press the link below for help",
   "addressValidationPartialOwnAddress": "Please enter the last 4 digits of your contact's account number, not your own. Press the link below for help",
   "helpFindAccount": "Help your contact find their account number",
-  "close": "Close"
+  "close": "Close",
+  "review": "Review"
 }
