+++ packages/mobile/locales/en-US/sendFlow7.json
@@ -105,6 +105,5 @@
     "body1": "Your contact can find their account number (also known as an address) at the top of the Settings menu.",
     "body2": "Account numbers begin with “0x” and are 42 characters long. They appear directly beneath the profile photo."
   },
-  "transferAddressChanged": "This contact changed their account number since the transaction. This payment was sent to:",
-  "review": "Review"
+  "transferAddressChanged": "This contact changed their account number since the transaction. This payment was sent to:"
 }
