+++ packages/mobile/locales/es-419/global.json
@@ -159,5 +159,6 @@
   "addressValidationFullOwnAddress": "Por favor, ingresa el número de cuenta de tu contacto, no el tuyo. Haz click en el siguiente enlace para obtener ayuda",
   "addressValidationPartialOwnAddress": "Por favor, ingresa los últimos 4 dígitos del número de cuenta de tu contacto, no el tuyo. Haz click en el siguiente enlace para obtener ayuda",
   "helpFindAccount": "Ayuda a tu contacto a encontrar su número de cuenta",
-  "close": "Cerrar"
+  "close": "Cerrar",
+  "review": "Revisión"
 }
