+++ packages/mobile/src/navigator/Headers.v2.tsx
@@ -1,9 +1,6 @@
-import QRCodeBorderlessIcon from '@celo/react-components/icons/QRCodeBorderless'
-import Times from '@celo/react-components/icons/Times'
 import colors from '@celo/react-components/styles/colors.v2'
 import fontStyles from '@celo/react-components/styles/fonts.v2'
 import { componentStyles } from '@celo/react-components/styles/styles'
-import { RouteProp } from '@react-navigation/core'
 import { StackNavigationOptions } from '@react-navigation/stack'
 import * as React from 'react'
 import { Trans } from 'react-i18next'
@@ -13,10 +10,6 @@ import CancelButton from 'src/components/CancelButton.v2'
 import CurrencyDisplay from 'src/components/CurrencyDisplay'
 import { CURRENCIES, CURRENCY_ENUM } from 'src/geth/consts'
 import i18n, { Namespaces } from 'src/i18n'
-import { navigate, navigateBack } from 'src/navigator/NavigationService'
-import { Screens } from 'src/navigator/Screens'
-import { TopBarIconButton, TopBarTextButton } from 'src/navigator/TopBarButton.v2'
-import { StackParamList } from 'src/navigator/types'
 import useSelector from 'src/redux/useSelector'
 import DisconnectBanner from 'src/shared/DisconnectBanner'
 
@@ -35,9 +28,6 @@ const styles = StyleSheet.create({
   header: {
     alignItems: 'center',
     justifyContent: 'center',
-  },
-  cancelBtn: {
-    color: colors.dark,
   },
 })
 
@@ -90,83 +80,6 @@ export const emptyHeader: StackNavigationOptions = {
       },
     }),
   },
-}
-
-export const exchangeTradeOptions = ({
-  route,
-}: {
-  route: RouteProp<StackParamList, Screens.ExchangeTradeScreen>
-}) => {
-  const { makerToken } = route.params?.makerTokenDisplay
-  const title =
-    makerToken === CURRENCY_ENUM.DOLLAR
-      ? i18n.t('exchangeFlow9:buyGold')
-      : i18n.t('exchangeFlow9:sellGold')
-  return {
-    ...headerWithCancelButton,
-    headerLeft: () => <CancelButton style={styles.cancelBtn} />,
-    headerTitle: () => <HeaderTitleWithBalance title={title} token={makerToken} />,
-  }
-}
-
-export const exchangeReviewOptions = ({
-  route,
-}: {
-  route: RouteProp<StackParamList, Screens.ExchangeReview>
-}) => {
-  const { makerToken } = route.params?.exchangeInput
-  const goExchangeHome = () => navigate(Screens.ExchangeHomeScreen)
-  const title =
-    makerToken === CURRENCY_ENUM.DOLLAR
-      ? i18n.t('exchangeFlow9:buyGold')
-      : i18n.t('exchangeFlow9:sellGold')
-  return {
-    ...headerWithCancelButton,
-    headerLeft: () => <CancelButton style={styles.cancelBtn} onCancel={goExchangeHome} />,
-    headerRight: () => (
-      <TopBarTextButton
-        title={i18n.t('global:edit')}
-        testID="EditButton"
-        onPress={navigateBack}
-        titleStyle={{ color: colors.goldDark }}
-      />
-    ),
-    headerTitle: () => <HeaderTitleWithBalance title={title} token={makerToken} />,
-  }
-}
-
-export const sendOptions = ({ route }: { route: RouteProp<StackParamList, Screens.Send> }) => {
-  const goQr = () => navigate(Screens.QRCode)
-  return {
-    ...emptyHeader,
-    headerLeft: () => <TopBarIconButton icon={<Times />} onPress={navigateBack} />,
-    headerLeftContainerStyle: { paddingLeft: 20 },
-    headerRight: () => (
-      <TopBarIconButton
-        icon={<QRCodeBorderlessIcon height={32} color={colors.greenUI} />}
-        onPress={goQr}
-      />
-    ),
-    headerRightContainerStyle: { paddingRight: 16 },
-    headerTitle: i18n.t(`sendFlow7:${route.params?.isRequest ? 'request' : 'send'}`),
-  }
-}
-
-export const sendAmountOptions = ({
-  route,
-}: {
-  route: RouteProp<StackParamList, Screens.SendAmount>
-}) => {
-  return {
-    ...emptyHeader,
-    headerLeft: () => <BackButton />,
-    headerTitle: () => (
-      <HeaderTitleWithBalance
-        title={i18n.t(`sendFlow7:${route.params?.isRequest ? 'request' : 'send'}`)}
-        token={CURRENCY_ENUM.DOLLAR}
-      />
-    ),
-  }
 }
 
 export const drawerHeader: StackNavigationOptions = {
